Д9: соблюдена структура jsx-файлов
