+++ src/components/genres/genres.jsx
@@ -3,27 +3,11 @@ import PropTypes from "prop-types";
 import {connect} from "react-redux";
 import {changeGenre, clearRenderedFilms} from "../../store/action";
 import {getFilteredFilms} from "../../store/reducers/app-state/selectors";
+import {getGenreInPlural, getFilterItemClass} from "../../utils";
 
 const Genres = (props) => {
   const {films, activeGenre, onGenreClick} = props;
   const genres = [`All genres`, ...new Set(films.map((film) => film.genre))];
-
-  const getFilterItem = (genre) => {
-    switch (genre) {
-      case `Comedy`:
-        return `Comedies`;
-      case `Drama`:
-        return `Dramas`;
-      case `Thriller`:
-        return `Thrillers`;
-      default:
-        return genre;
-    }
-  };
-
-  const getFilterItemClass = (genre) => {
-    return activeGenre === genre ? `catalog__genres-item catalog__genres-item--active` : `catalog__genres-item`;
-  };
 
   return (
     <ul className="catalog__genres-list">
@@ -31,7 +15,7 @@ const Genres = (props) => {
         return (
           <li
             key={genre + i}
-            className={getFilterItemClass(genre)}
+            className={getFilterItemClass(genre, activeGenre)}
           >
             <a
               href="#"
@@ -40,12 +24,24 @@ const Genres = (props) => {
                 evt.preventDefault();
                 onGenreClick(genre);
               }}
-            >{getFilterItem(genre)}</a>
+            >{getGenreInPlural(genre)}</a>
           </li>
         );
       })}
     </ul>
   );
+};
+
+Genres.propTypes = {
+  films: PropTypes.arrayOf(PropTypes.shape({
+    id: PropTypes.number.isRequired,
+    name: PropTypes.string.isRequired,
+    previewImage: PropTypes.string.isRequired,
+    previewVideoLink: PropTypes.string.isRequired,
+    genre: PropTypes.string.isRequired,
+  })).isRequired,
+  activeGenre: PropTypes.string.isRequired,
+  onGenreClick: PropTypes.func.isRequired,
 };
 
 const mapStateToProps = ({APP_STATE}) => ({
@@ -60,18 +56,6 @@ const mapDispatchToProps = (dispatch) => ({
     dispatch(clearRenderedFilms());
   }
 });
-
-Genres.propTypes = {
-  films: PropTypes.arrayOf(PropTypes.shape({
-    id: PropTypes.number.isRequired,
-    name: PropTypes.string.isRequired,
-    previewImage: PropTypes.string.isRequired,
-    previewVideoLink: PropTypes.string.isRequired,
-    genre: PropTypes.string.isRequired,
-  })).isRequired,
-  activeGenre: PropTypes.string.isRequired,
-  onGenreClick: PropTypes.func.isRequired,
-};
 
 export {Genres};
 export default connect(mapStateToProps, mapDispatchToProps)(Genres);
