+++ src/const.js
@@ -11,6 +11,8 @@ export const RatingMarkType = {
   AWESOME: `Awesome`,
 };
 
+export const RATING_MARKS = [1, 2, 3, 4, 5];
+
 export const FilmPageNav = {
   OVERVIEW: `Overview`,
   DETAILS: `Details`,
